Add files via upload
So this is basically just a little thing where I was trying to combine my Chinese language skills with my Java skills. The only issue I had so far with it is that when the user uses Chinese as input and then displays what the user entered it does not display it correctly and I wasn't sure how to fix that. My best guess for fixing that issue is importing an language pack or something other than that I do not know.